Fix warnings caused by -Wall

+++ src/Spectrogram/Audio/System/Blocking.cpp
@@ -12,7 +12,7 @@ Spectrogram::Audio::System::Blocking::Blocking(std::unique_ptr<Backend::Backend>
 
 void Spectrogram::Audio::System::Blocking::start(const Spectrogram::Audio::Device &device, size_t frames) {
 
-    for (int channel = 0; channel < device.channelCount; ++channel) {
+    for (size_t channel = 0; channel < device.channelCount; ++channel) {
         _channelQueues.emplace_back(frames * 2);
     }
     _frames = frames;
@@ -43,7 +43,7 @@ Spectrogram::Audio::Buffer Spectrogram::Audio::System::Blocking::getBuffer() {
 
 void Spectrogram::Audio::System::Blocking::pushSamples(const std::vector<Sample *> &arrays, size_t length) {
 
-    for (int channel = 0; channel < _channelQueues.size(); ++channel) {
+    for (size_t channel = 0; channel < _channelQueues.size(); ++channel) {
         auto samplesPushed = _channelQueues[channel].push(arrays[channel], length);
         assert(samplesPushed == length);
     }
